requirements.txt
faiss-cpu
faiss-gpu
numpy
pickle-mixin
typing-extensions